revert 03e0257759356c5f0cc748d1cf4511c3e7254d48 (something's fishy)

+++ include/cpu/cpu.h
@@ -22,7 +22,7 @@
 #include "cpu/iset.h"
 
 extern int P;
-extern uint32_t N;
+extern int16_t N;
 extern int cpu_mod;
 
 #ifdef WITH_DEBUGGER
