+++ include/mem/mem.h
@@ -50,8 +50,8 @@ int mem_cpu_get(int nb, uint16_t addr, uint16_t *data);
 int mem_cpu_put(int nb, uint16_t addr, uint16_t data);
 int mem_cpu_mget(int nb, uint16_t saddr, uint16_t *dest, int count);
 int mem_cpu_mput(int nb, uint16_t saddr, uint16_t *src, int count);
-int mem_get_byte(int nb, uint32_t addr, uint8_t *data);
-int mem_put_byte(int nb, uint32_t addr, uint8_t data);
+int mem_get_byte(int nb, uint16_t addr, uint8_t *data);
+int mem_put_byte(int nb, uint16_t addr, uint8_t data);
 
 void mem_clear();
 int mem_seg_load(FILE *f, uint16_t *ptr);
